Issue #10, Move User class to model folder and Create its wraper
