+++ app/src/main/java/com/rest/model/UserWraper.java
@@ -1,0 +1,12 @@
+package com.rest.model;
+
+/**
+ * Created by ehab on 2/4/15.
+ */
+public class UserWraper {
+    private User user;
+
+    public UserWraper(User u) {
+        this.user = u;
+    }
+}
